drawer, fix login and signup styles, options button on card

+++ website/src/components/Item.tsx
@@ -1,24 +1,34 @@
-import React, { PropsWithChildren, Fragment, useContext } from 'react';
+import React, {
+  PropsWithChildren,
+  Fragment,
+  useContext,
+  useState
+} from 'react';
 import { Card, StyledBody, StyledAction, StyledThumbnail } from 'baseui/card';
+import { StatefulMenu } from 'baseui/menu';
 
 import { Button } from 'baseui/button';
-
+import { StatefulPopover, PLACEMENT } from 'baseui/popover';
 import EditInPlace from './EditInPlace';
 
 import { ItemInterface } from '../context/types';
 import { useStyletron } from 'baseui';
+import { Overflow } from 'baseui/icon';
 
 import ItemsContext from '../context/items/ItemsContext';
+import { Block } from 'baseui/block';
 
 // https://source.unsplash.com/user/erondu/300x300
 
-export default function Item(props: PropsWithChildren<ItemInterface>) {
-  const { title, body, id } = props;
-  const [css] = useStyletron();
+export default function Item(props: any) {
+  const { title, body, id, openMenu } = props;
+  const [open, setOpen] = useState(false);
   const { deleteItem } = useContext(ItemsContext);
-
-  function handleDelete() {
-    deleteItem(id);
+  const [css] = useStyletron();
+  function handleDelete(label: string) {
+    if (label === 'Delete') {
+      deleteItem(id);
+    }
   }
   return (
     <Fragment>
@@ -34,12 +44,38 @@ export default function Item(props: PropsWithChildren<ItemInterface>) {
               justifyContent: 'flex-end'
             })}
           >
-            <Button $as="a" href={body} size="compact" target="_blank">
+            <Button
+              $as="a"
+              href={body}
+              kind="primary"
+              size="compact"
+              target="_blank"
+            >
               Visit
             </Button>
-            <Button size="compact" kind="minimal" onClick={handleDelete}>
-              Delete
-            </Button>
+            <StatefulPopover
+              placement={PLACEMENT.bottomLeft}
+              content={({ close }) => (
+                <StatefulMenu
+                  items={[
+                    { label: 'Share' },
+                    { label: 'Favorite' },
+                    { label: 'Delete' }
+                  ]}
+                  onItemSelect={({ item: { label } }) => {
+                    close();
+                    handleDelete(label);
+                  }}
+                  // overrides={{
+                  //   List: { style: { height: '150px', width: '138px' } }
+                  // }}
+                />
+              )}
+            >
+              <Button kind="tertiary" size="compact">
+                <Overflow />
+              </Button>
+            </StatefulPopover>
           </div>
         </StyledAction>
       </Card>
